+++ comm.c
@@ -948,9 +948,9 @@ int get_max_players(void)
     if (limit.rlim_max == RLIM_INFINITY)
       max_descs = CONFIG_MAX_PLAYING + NUM_RESERVED_DESCS;
     else
-      max_descs = MIN(CONFIG_MAX_PLAYING + NUM_RESERVED_DESCS, limit.rlim_max);
+      max_descs = MIN(CONFIG_MAX_PLAYING + NUM_RESERVED_DESCS, (int)limit.rlim_max);
 #else
-  max_descs = MIN(CONFIG_MAX_PLAYING + NUM_RESERVED_DESCS, limit.rlim_max);
+  max_descs = MIN(CONFIG_MAX_PLAYING + NUM_RESERVED_DESCS, (int)limit.rlim_max);
 #endif
   }
